# WebAPI/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Web API ",
      "description": "API del sistema de capacitación",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
      "http",
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/api/getUsuario/{id}": {
        "get": {
          "tags": [
            "Usuario"
          ],
          "summary": "Encuentra un usuario por id",
          "description": "Regresa el usuario con el id solicitado",
          "operationId": "getUsuario",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Id del usuario",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          }
        }
      },
      "/api/addUsuario": {
        "post": {
          "tags": [
            "Usuario"
          ],
          "summary": "Agrega un nuevo Usuario",
          "description": "Agrega un Usaurio a la base de datos",
          "operationId": "addUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Nuevo usuario",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "405": {
              "description": "Datos inválidos"
            }
          }
        }
      },
      "/api/updateUsuario/{id}": {
        "put": {
          "tags": [
            "Usuario"
          ],
          "summary": "Actualiza un usuario",
          "description": "Actualiza un usuario a la base de datos",
          "operationId": "updateUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Id del usuario a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Datos nuevos",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "400": {
              "description": "Usuario Inválido"
            },
            "404": {
              "description": "Usuario no encontrado"
            }
          }
        }
      },
      "/api/updateUsuarioAdmin/{id}": {
        "put": {
          "tags": [
            "Usuario"
          ],
          "summary": "Actualiza un usuario",
          "description": "Actualiza un usuario a la base de datos",
          "operationId": "updateUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Id del usuario a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Datos nuevos",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "400": {
              "description": "Usuario Inválido"
            },
            "404": {
              "description": "Usuario no encontrado"
            }
          }
        }
      },
      "/api/deleteUsuario/{id}": {
        "delete": {
          "tags": [
            "Usuario"
          ],
          "summary": "Borra un Usuario",
          "description": "Borra un Usuario de la base de datos",
          "operationId": "deleteUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Id del Usuario a borrar",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "400": {
              "description": "Usuario Inválido"
            },
            "404": {
              "description": "Usuario no encontrado"
            }
          }
        }
      },
...
    },
    "definitions": {
      "Usuario": {
        "type": "object",
        "properties": {
          "idUsuario": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "apellidoPaterno": {
            "type": "string"
          },
          "apellidoMaterno": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "tipoUsuario": {
            "type": "integer",
            "format": "int64"
          },
          "puesto": {
            "type": "string"
          },
          "idAdministrador": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }